fix: report QA-04 as NeedsReview when Security Insights is absent
